refactor(store/db): move `select_account_vault_at_block()` to tests
